Add choice apache/nginx and self reload

+++ vhost.sh
@@ -5,27 +5,69 @@ nginxSuffix=nginx
 
 read -p "vHost name (domain): " serverName
 read -p "Alias name: " aliasName
-read -p "Document root (sans /var/www/): " documentRoot
+read -p "Document root (without /var/www/): " documentRoot
 documentRoot="/var/www/"$documentRoot
 
-defaultApache=/etc/apache2/site-available/000-default-script.$apacheSuffix
-vHostApache=/etc/apache2/site-available/$serverName.$apacheSuffix
 
-cp $defaultApache $vHostApache
-sed -i 's/${serverName}/'$serverName'/' $vHostApache
-sed -i 's/${aliasName}/'$aliasName'/' $vHostApache
-sed -i 's/${documentRoot}/'$documentRoot'/' $vHostApache
-sudo a2enconf $serverName
-sudo systemctl restart apache2
 
-defaultNginx=/etc/nginx/site-available/000-default-script.$nginxSuffix
-vHostNginx=/etc/nginx/site-available/$serverName.$nginxSuffix
-vHostEnabledNginx=/etc/nginx/site-enabled/$serverName.$nginxSuffix
+while true; do
+      read -p "Create on Apache  ? " doApache
+      case $doApache in
+          [Yy] ) doApache="y" break;;
+          [Nn] ) doApache="n" break;;
+          * ) echo "Yes or No";;
+      esac
+done
 
-cp $defaultNginx $vHostNginx
-sed -i 's/${serverName}/'$serverName'/' $vHostNginx
-sed -i 's/${aliasName}/'$aliasName'/' $vHostNginx
-sed -i 's~${documentRoot}~'$documentRoot'~' $vHostNginx
-sudo ln -s $vHostNginx $vHostEnabledNginx
-sudo systemctl restart nginx
-echo "vHost created"
+if [ $doApache == "y" ]; then
+
+    defaultApache=/etc/apache2/site-available/000-default-script.$apacheSuffix
+    vHostApache=/etc/apache2/site-available/$serverName.$apacheSuffix
+
+    cp $defaultApache $vHostApache
+    sed -i 's/${serverName}/'$serverName'/' $vHostApache
+    sed -i 's/${aliasName}/'$aliasName'/' $vHostApache
+    sed -i 's~${documentRoot}~'$documentRoot'~' $vHostApache
+    sudo a2enconf $serverName
+    sudo systemctl restart apache2
+    echo "vHost created and enabled on Apache"
+fi
+
+while true; do
+      read -p "Create on Nginx  ? " doNginx
+      case $doNginx in
+          [Yy] ) doNginx="y" break;;
+          [Nn] ) doNginx="n" break;;
+          * ) echo "Yes or No";;
+      esac
+done
+
+if [ $doNginx == "y" ]; then
+    defaultNginx=/etc/nginx/site-available/000-default-script.$nginxSuffix
+    vHostNginx=/etc/nginx/site-available/$serverName.$nginxSuffix
+    vHostEnabledNginx=/etc/nginx/site-enabled/$serverName.$nginxSuffix
+
+    cp $defaultNginx $vHostNginx
+    sed -i 's/${serverName}/'$serverName'/' $vHostNginx
+    sed -i 's/${aliasName}/'$aliasName'/' $vHostNginx
+    sed -i 's~${documentRoot}~'$documentRoot'~' $vHostNginx
+    sudo ln -s $vHostNginx $vHostEnabledNginx
+    sudo systemctl restart nginx
+    echo "vHost created and enabled on NGINX"
+fi
+
+while true; do
+      read -p "Create more vHost ? " moreVHost
+      case $moreVHost in
+          [Yy] ) moreVHost="y" break;;
+          [Nn] ) moreVHost="n" break;;
+          * ) echo "Yes or No";;
+      esac
+done
+
+if [ $moreVHost == "y" ]; then
+    addMoreVHost=$(readlink -f "$0")
+    exec $addMoreVHost
+else
+  exit
+fi
